SunOS JDKs need to be swapped for "sparc" and "x86"
Currently, the JDKs for SunOS are extracted into the wrong directories.
The "x86" tarball is being extracted into the "sparc" directory, and the
"sparc" tarball into the "x86" directory. As a result, the product fails
when using these artifacts, and interacting with SunOS systems. This
change fixes this problem, such that the correct tarball is extracted
into the correct architecture specific directory.